proje4/requirements.txt
torch>=2.0
torchvision
timm>=1.0
albumentations>=1.4
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
Pillow>=10.0
tqdm>=4.65
tensorboard
grad-cam>=1.5
